benchmarks/requirements.txt
pytest-benchmark
tenacity
